Add: receiver interface & Fix: functions scope
